requirements.txt
streamlit
PyMuPDF
google-generativeai
langchain
chromadb
beautifulsoup4
pyyaml
